back office terminer, travail sur la partie actualite

+++ database/migrations/2025_05_21_145842_create_critiques_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id("id_critique");
             $table->foreignId('id_movie')->constrained('movies', 'id_movie')->onDelete('cascade');
             $table->foreignId('id_user')->constrained('users','user_id');
-            $table->decimal('note',4,2);
+            $table->integer('note');
             $table->text("critique");
             $table->integer("nbr_like");
             $table->timestamps();
